+++ NEWS.md
@@ -1,6 +1,6 @@
-# dlookr 0.6.2.9000
+# dlookr 0.6.2.9001
 
-## MINOR CHANGES
+## MAJOR CHANGES
 
 * Add group_by() support functions. (thanks to Jüri Kuusik, #90)
    - diagnose()
@@ -8,7 +8,11 @@
    - diagnose_numeric()
    - diagnose_outlier()
   
-  
+* Add group_by() support functions for DBMS. (#93)
+   - diagnose.tbl_dbi()
+   - diagnose_category.tbl_dbi()
+
+
 
 # dlookr 0.6.1